requirements.txt
asgiref==3.3.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
cycler==0.10.0
Django==3.2.4
django-crispy-forms==1.11.2
django-embed-video==1.4.0
django-multiselectfield==0.1.12
gunicorn==20.1.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
matplotlylib==0.1.0
mutagen==1.45.1
numpy==1.19.5
Pillow==8.1.0
plotly==4.14.1
pygame==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
requests==2.25.1
retrying==1.3.3
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
webdriver-manager==3.3.0
whitenoise==5.2.0
